Assuming the new lazy load RawReader as the default one.
